Bug Fix: default_from overriding `from` in emails sent via APIMailer
Small bug fix I noticed earlier today. 😄

When using the APIMailer with ActionMailer.

If you have `default_from` set, but then set a different `from` for the email. It currently overwrites the `from` with the default value.

Making it impossible to send email from an address other than the `default_from`.

This change checks to see if `from` is set before using the value set in `default_from`.